cleanup track update trigger migration
