--- discovery-provider/alembic/trigger_sql/drop_triggers.sql
@@ -1,1 +1,0 @@
-drop trigger on_track on tracks;
